vendimia/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.6
django-cacheops==4.0.1
django-celery==3.2.1
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-redis==4.8.0
foundationform==0.2
funcy==1.9.1
idna==2.6
kombu==3.0.37
olefile==0.44
Pillow==4.3.0
pytz==2017.2
redis==2.10.6
requests==2.18.4
six==1.11.0
urllib3==1.22
vine==1.1.4
